Se termina la verificación del blockchain

+++ lib/models/miner_model.dart
@@ -19,7 +19,7 @@ class MinerModel {
         index: json['index'],
         name: json['name'],
         blocksMined: json['blocksMined'],
-        totalCoins: json['totalCoins'],
+        totalCoins: json['totalCoins'].toDouble(),
         work: json['work'],
         workPercent: json['workPercent'],
       );
